Fix erratic behaviour with occasional off-by-one-bit error.
Turns out that the TLC5940 needs one extra SCLK pulse after latching GS data
to correctly put the status information into the shift register (this is
actually mentioned in the data sheet).

With this extra pulse, there is no need to switch to different SCLK polarity
for reading back DC values, and operations seems to be now stable at 21 MHz.