Defer slow part of per-scanplane interrupt to a lower prio software interrupt.
The per-scanplane interrupt starts with timing-critical latch of new
data on the TLC5955's. But after that comes generation of next
scanplanes, which is a long-running operation that is not timing
critical.

So run the latching in a high-priority interrupt, and defer the
scanplane generation to a lower-priority software interrupt. This way,
other, medium-priority interrupts can run without being delayed by
scanplane generation.