First working nRF async packet reception.
The nRF24L01+ module is configured in receive mode.

A GPIO interrupt triggers whenever a packet is received. An interrupt-driven
state machine runs to fetch packets from the nRF and deliver them to a
callback.

A debug command "reset to bootloader" is handled in the callback, allowing
to wireless flash without manually resetting to the bootloader. More
command handling to be added later.

Code ported from lm4f_pov.c.