Improved battery voltage monitoring.
Do an extra ADC measurement on in the internal 1.2V reference, and use
that to adjust the voltage measurement of the battery voltage.

This way, we should get correct readings when the battery voltage
drops low and the LDO is no longer able to maintain a 3.3V reference /
supply voltage. This seems good, as low battery voltage is arguably
the most interesting case for battery voltage monitoring.